CRUD Cards, Schemas and relations finished

+++ src/controllers/albums.js
@@ -3,30 +3,35 @@ import crypto from 'crypto'
 
 export class AlbumController {
   static async getMyAlbums(req, res) {
-    // Temporalmente sin pedir userId (viene de un middleware com req.userId mas adelante)
-    const response = await pool.query('SELECT * FROM album_cards') //where user_id = $1, [req.userId]
-    await pool.end()
+    const userId = req.userId
+    const response = await pool.query(
+      'SELECT * FROM album_cards WHERE user_id = $1',
+      [userId]
+    ) 
     res.status(200).json(response.rows)
   }
 
   // Creates a new Album with a randomUUID
   static async createNewAlbum(req, res) {
+    const userId = req.userId
     const { name } = req.body
     const randomUuid = crypto.randomUUID()
-    const textQuery = 'INSERT INTO album_cards(album_id, name) VALUES($1, $2)'
-    await pool.query(textQuery, [randomUuid, name])
+    await pool.query(
+      'INSERT INTO album_cards(album_id, name, user_id) VALUES($1, $2, $3)',
+      [randomUuid, name, userId]
+    )
 
     res.status(201).send('Album creado correctamente')
   }
 
   static async editAlbumName(req, res) {
+    const userId = req.userId
     const { albumId } = req.params
     const { newName } = req.body
     const result = await pool.query(
-      'UPDATE album_cards SET name = $1 WHERE album_id = $2 RETURNING *',
-      [newName, albumId]
+      'UPDATE album_cards SET name = $1 WHERE album_id = $2 AND user_id = $3 RETURNING *',
+      [newName, albumId, userId]
     )
-    await pool.end()
     if (!result.rowCount) {
       return res.status(404).json({ message: 'El album no existe' })
     }
@@ -35,15 +40,16 @@ export class AlbumController {
   }
 
   static async deleteAlbum(req, res) {
+    const userId = req.userId
     const { albumId } = req.params
     const result = await pool.query(
-      'DELETE FROM album_cards WHERE album_id = $1 RETURNING *',
-      [albumId]
+      'DELETE FROM album_cards WHERE album_id = $1 AND user_id = $2 RETURNING *',
+      [albumId, userId]
     )
-    await pool.end()
     console.log(result)
+    
     if (!result.rowCount) {
-      return res.status(404).json({ message: 'El album no existe' })
+      return res.status(404).json({ message: 'El album no existe o no te pertenece' })
     }
     res.sendStatus(204)
   }
